assetprot/sportsentinel-ml/requirements.txt
google-cloud-storage
google-cloud-firestore
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-vision
google-cloud-speech
google-generativeai
faiss-cpu
opencv-python
moviepy
librosa
torch
torchvision
transformers
Pillow
numpy
pandas
scipy
scikit-learn
python-dotenv
requests
tqdm
soundfile
